WIP community page reworked in some womponents
